superagent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.44.0
crewai==1.14.1
langchain==0.3.27
langchain-groq==0.2.0
langchain-community==0.3.27
python-dotenv==1.1.1
pydantic==2.11.10
httpx==0.28.1
cachetools==5.5.0
python-multipart==0.0.22
